move MultiStringOverride.cs into auto faker helpers
