+++ backend/FwLite/FwLiteProjectSync.Tests/EntrySyncTests.cs
@@ -3,7 +3,6 @@ using FwLiteProjectSync.Tests.Fixtures;
 using MiniLcm.Models;
 using MiniLcm.SyncHelpers;
 using MiniLcm.Tests.AutoFakerHelpers;
-using MiniLcm.Tests.FakerOverrids;
 using Soenneker.Utils.AutoBogus;
 
 namespace FwLiteProjectSync.Tests;
