+++ backend/FwLite/LcmCrdt.Tests/DataModelSnapshotTests.cs
@@ -7,7 +7,7 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
-using MiniLcm.Tests.FakerOverrids;
+using MiniLcm.Tests.AutoFakerHelpers;
 using SIL.Harmony.Changes;
 using SIL.Harmony.Entities;
 using Soenneker.Utils.AutoBogus;
