+++ backend/FwLite/MiniLcm.Tests/MiniLcmTestBase.cs
@@ -1,5 +1,4 @@
 ﻿using MiniLcm.Tests.AutoFakerHelpers;
-using MiniLcm.Tests.FakerOverrids;
 using Soenneker.Utils.AutoBogus;
 
 namespace MiniLcm.Tests;
